Created EditStudentPortfolioModal and refactor code.
