+++ src/modals/EditStudentPortfoliModal/EditStudentPortfoliModal.styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const ModalContainer = styled.div`
+   position: absolute;
+   top: 0;
+   left: 0;
+   width: 100%;
+`
